src/requirements.txt
Django==1.7.1
argparse==1.2.1
dj-static==0.0.6
static3==0.5.1
wsgiref==0.1.2

#rest
djangorestframework
markdown
django-filter

#debuger
Werkzeug